Re-order examples and add handleClick
