+++ 13-form-validation/a-validation/src/Field.css
@@ -1,4 +1,4 @@
-// App.css
 input:invalid + .pure-form-message-inline {
   color: #c61c06;
 }
+
